support prepared stmt
